fix: address /code-review 331 findings on Graph API versions and OAuth error surfacing
Unify Facebook/Instagram Graph API version defaults behind a shared
api/src/lib/graph-api-version.util.ts helper (fixes the v23.0-vs-v25.0
mismatch between auth-url.helper.ts and the rest of api/, and dedupes
instagram.service.ts's hardcoded refresh-token URL), guard against
trailing slashes on operator-configurable Graph API base URLs in
trigger/ and dashboard/, and extract the duplicated OAuth
error-parsing/response logic in both dashboard callback loaders into
oauth-callback-response.ts — sanitizing literal "|" out of
provider-supplied error text and skipping DB lookups that are only
needed on the success path.
