+++ dashboard/app/lib/.server/social-accounts/social-account.constants.ts
@@ -1,11 +1,13 @@
 export const SOCIAL_ACCOUNT_PHOTO_BUCKET_NAME = "social-account-photos";
 export const REDIRECT_APP_URL = process.env.APP_URL || "http://localhost:5173";
 
-export const FACEBOOK_GRAPH_API_URL =
-  process.env.FACEBOOK_GRAPH_API_URL || "https://graph.facebook.com";
+export const FACEBOOK_GRAPH_API_URL = (
+  process.env.FACEBOOK_GRAPH_API_URL || "https://graph.facebook.com"
+).replace(/\/+$/, "");
 export const FACEBOOK_API_VERSION =
   process.env.FACEBOOK_API_VERSION || "v25.0";
-export const INSTAGRAM_GRAPH_API_URL =
-  process.env.INSTAGRAM_GRAPH_API_URL || "https://graph.instagram.com";
+export const INSTAGRAM_GRAPH_API_URL = (
+  process.env.INSTAGRAM_GRAPH_API_URL || "https://graph.instagram.com"
+).replace(/\/+$/, "");
 export const INSTAGRAM_API_VERSION =
   process.env.INSTAGRAM_API_VERSION || "v23.0";
