+++ dashboard/app/routes/callback.$provider.account/route.loader.ts
@@ -1,7 +1,10 @@
-import { redirect } from "react-router";
 import { addSocialAccountConnections } from "~/lib/.server/social-accounts/social-account";
 import { withSupabase } from "~/lib/.server/supabase";
 import type { Database } from "~/lib/.server/database.types";
+import {
+  parseOauthCallbackError,
+  createOauthCallbackResponse,
+} from "~/lib/.server/social-accounts/oauth-callback-response";
 
 type SocialProviderEnum = Database["public"]["Enums"]["social_provider"];
 
@@ -17,31 +20,14 @@ export const loader = withSupabase(async function ({
 
   let { provider } = params;
 
-  const oauthError = url.searchParams.get("error");
-  const oauthErrorReason = url.searchParams.get("error_reason");
-  const oauthErrorDescription = url.searchParams.get("error_description");
-  const oauthErrorMessage = oauthError
-    ? oauthErrorDescription || oauthErrorReason || oauthError
-    : null;
-
-  if (oauthError) {
-    console.error(
-      "OAuth provider returned an error during account connection",
-      {
-        provider,
-        error: oauthError,
-        error_reason: oauthErrorReason,
-        error_description: oauthErrorDescription,
-      },
-    );
-  }
+  const { oauthErrorMessage } = parseOauthCallbackError(url, provider);
 
   const key =
     (url.searchParams.get("oauth_token") as string) ||
     (url.searchParams.get("state") as string);
 
   if (!key) {
-    return createResponse({
+    return createOauthCallbackResponse({
       isSuccess: false,
       errors: [oauthErrorMessage || "Auth state not set"],
       isLoggedIn,
@@ -68,26 +54,11 @@ export const loader = withSupabase(async function ({
   )?.value;
 
   if (!projectId || !provider) {
-    return createResponse({
+    return createOauthCallbackResponse({
       isSuccess: false,
       errors: [oauthErrorMessage || "Project Id or Provider not found"],
       isLoggedIn,
     });
-  }
-
-  const connectionType = oauthData.data?.find(
-    (d) => d.key === "connection_type",
-  )?.value;
-  if (
-    connectionType &&
-    provider === "instagram" &&
-    connectionType === "facebook"
-  ) {
-    provider = "instagram_w_facebook";
-  }
-
-  if (provider === "x" && connectionType === "oauth2") {
-    provider = "x_oauth2";
   }
 
   const normalizedProvider =
@@ -116,7 +87,7 @@ export const loader = withSupabase(async function ({
 
   if (projectError || !project) {
     console.error("Project not found");
-    return createResponse({
+    return createOauthCallbackResponse({
       isSuccess: false,
       errors: [oauthErrorMessage || "Project not found"],
       projectId,
@@ -125,12 +96,8 @@ export const loader = withSupabase(async function ({
     });
   }
 
-  const providerAppCredentials = project.social_provider_app_credentials.find(
-    (appCredential) => appCredential.provider === provider,
-  );
-
   if (oauthErrorMessage) {
-    return createResponse({
+    return createOauthCallbackResponse({
       isSuccess: false,
       teamId: project.team_id,
       projectId,
@@ -141,9 +108,28 @@ export const loader = withSupabase(async function ({
     });
   }
 
+  const connectionType = oauthData.data?.find(
+    (d) => d.key === "connection_type",
+  )?.value;
+  if (
+    connectionType &&
+    provider === "instagram" &&
+    connectionType === "facebook"
+  ) {
+    provider = "instagram_w_facebook";
+  }
+
+  if (provider === "x" && connectionType === "oauth2") {
+    provider = "x_oauth2";
+  }
+
+  const providerAppCredentials = project.social_provider_app_credentials.find(
+    (appCredential) => appCredential.provider === provider,
+  );
+
   if (!providerAppCredentials && provider !== "bluesky") {
     console.error("Provider app credentials not found for project");
-    return createResponse({
+    return createOauthCallbackResponse({
       projectId,
       provider: normalizedProvider,
       teamId: project.team_id,
@@ -174,7 +160,7 @@ export const loader = withSupabase(async function ({
 
     if (successConnections.length === 0) {
       errors.push("No valid accounts found");
-      return createResponse({
+      return createOauthCallbackResponse({
         isSuccess: false,
         teamId: project.team_id,
         projectId,
@@ -186,7 +172,7 @@ export const loader = withSupabase(async function ({
       });
     }
 
-    return createResponse({
+    return createOauthCallbackResponse({
       isSuccess: true,
       teamId: project.team_id,
       projectId,
@@ -199,7 +185,7 @@ export const loader = withSupabase(async function ({
     });
   } catch (error) {
     console.error(error);
-    return createResponse({
+    return createOauthCallbackResponse({
       isSuccess: false,
       errors: [
         (error as { message?: string })?.message ||
@@ -213,57 +199,3 @@ export const loader = withSupabase(async function ({
     });
   }
 });
-
-//Either return to component or redirect to project callback url
-const createResponse = ({
-  teamId,
-  projectId,
-  provider,
-  isSuccess,
-  callbackUrl,
-  accountIds,
-  failedAccountIds,
-  errors,
-  isLoggedIn,
-}: {
-  teamId?: string;
-  projectId?: string;
-  provider?: string;
-  isSuccess: boolean;
-  accountIds?: string[];
-  failedAccountIds?: string[];
-  errors?: string[];
-  callbackUrl?: string | null | undefined;
-  isLoggedIn?: boolean;
-}) => {
-  const error = errors && errors.length > 0 ? errors.join("|") : null;
-
-  if (callbackUrl) {
-    const authParams = new URLSearchParams([
-      ["provider", provider || ""],
-      ["projectId", projectId || ""],
-      ["isSuccess", isSuccess ? "true" : "false"],
-      ["accountIds", accountIds?.join(",") || ""],
-    ]);
-
-    if (failedAccountIds && failedAccountIds.length > 0) {
-      authParams.append("failedAccountIds", failedAccountIds?.join(","));
-    }
-    if (error) {
-      authParams.append("error", error);
-    }
-
-    return redirect(`${callbackUrl}?${authParams.toString()}`);
-  }
-
-  return {
-    teamId,
-    projectId,
-    provider,
-    isSuccess,
-    accountIds,
-    failedAccountIds,
-    error,
-    isLoggedIn,
-  };
-};
